refactor: extract requireSameOrigin and invitationRateLimit to middle/
Move requireSameOrigin to src/server/middle/requireSameOrigin.ts and
invitationRateLimit to src/server/middle/invitationRateLimit.ts, following
the requireUser.ts pattern. Update invitationController.ts to import both
from middle/ instead of defining them inline.

Closes architecture-review finding 9c7.16.
