+++ src/server/controllers/invitationController.ts
@@ -1,5 +1,5 @@
 /**
- * invitationController.ts — Admin invitation create route + anonymous lookup/accept + rate limiter
+ * invitationController.ts — Admin invitation create route + anonymous lookup/accept
  *
  * Spec: specs/002-invitation-system/spec.md §FR-001..FR-012, §FR-013..FR-019, §FR-020
  * Plan: plan.md §Security Considerations (Pass 1/2/3 rate-limit, Pass 4/6 CSRF,
@@ -33,205 +33,9 @@ import {
   DecryptError,
 } from "../auth/invitationStore";
 import secrets from "../util/secrets";
-import { DEFAULT_BASE_URL, getTrustedOrigins } from "../auth/trustedOrigins";
-
-// ---------------------------------------------------------------------------
-// Shared Origin / CSRF middleware (plan.md Pass 4/6)
-//
-// Allow-list logic mirrors auth.ts `trustedOrigins` exactly so they cannot drift:
-//   - BETTER_AUTH_URL set → allow only that origin
-//   - NODE_ENV=development (and BETTER_AUTH_URL unset) → allow dev ports
-//   - NODE_ENV=test (and BETTER_AUTH_ENFORCE_ORIGIN !== '1') → skip check
-//   - otherwise → deny
-//
-// Check Origin first; fall back to Referer; reject 403 if neither matches
-// or neither is present.
-// ---------------------------------------------------------------------------
-
-/**
- * Returns the allowed origins for the current environment.
- *
- * Delegates to getTrustedOrigins() from trustedOrigins.ts so this middleware
- * and auth.ts's betterAuth trustedOrigins config can never drift apart
- * (architecture-review remediation 9c7.13; plan.md Pass 6).
- */
-function getAllowedOrigins(): string[] | null {
-  // Production with no BETTER_AUTH_URL → allow nothing (secrets.ts already
-  // throws at startup if BETTER_AUTH_URL is absent in production)
-  return getTrustedOrigins();
-}
-
-/**
- * Express middleware: validates the Origin (or Referer) header against the
- * BETTER_AUTH_URL allow-list for state-changing invitation routes.
- *
- * Skipped when NODE_ENV=test AND BETTER_AUTH_ENFORCE_ORIGIN !== '1', so the
- * normal Jest suite can exercise the routes without spoofing a browser Origin.
- * The integration test server sets BETTER_AUTH_ENFORCE_ORIGIN=1 to exercise
- * the live CSRF check (plan.md Pass 4/6).
- */
-export function requireSameOrigin(req: Request, res: Response, next: NextFunction): void {
-  // Skip in test mode unless integration flag is set
-  if (process.env.NODE_ENV === "test" && process.env.BETTER_AUTH_ENFORCE_ORIGIN !== "1") {
-    next();
-    return;
-  }
-
-  const allowedOrigins = getAllowedOrigins();
-
-  // No allowed origins configured → reject all (production misconfiguration
-  // is caught at startup by secrets.ts; this is a belt-and-suspenders guard)
-  if (!allowedOrigins || allowedOrigins.length === 0) {
-    res.status(403).json({ error: "Forbidden" });
-    return;
-  }
-
-  // Prefer Origin header; fall back to Referer
-  const origin = req.headers.origin;
-  let candidate: string | undefined;
-
-  if (origin) {
-    candidate = origin;
-  } else {
-    const referer = req.headers.referer;
-    if (referer) {
-      try {
-        const refererUrl = new URL(referer);
-        candidate = refererUrl.origin;
-      } catch {
-        // Malformed Referer → treat as absent
-      }
-    }
-  }
-
-  if (!candidate) {
-    res.status(403).json({ error: "Forbidden" });
-    return;
-  }
-
-  // Strip trailing slash from candidate for comparison
-  const normalizedCandidate = candidate.replace(/\/$/, "");
-  const allowed = allowedOrigins.some((o) => o.replace(/\/$/, "") === normalizedCandidate);
-
-  if (!allowed) {
-    res.status(403).json({ error: "Forbidden" });
-    return;
-  }
-
-  next();
-}
-
-// ---------------------------------------------------------------------------
-// Rate limiter middleware (plan.md Pass 1/2/3/12)
-//
-// Stores per-IP request counts in the invitationRateLimit table:
-//   key text PK, count int, lastRequest bigint (milliseconds since epoch)
-//
-// Atomic UPSERT with window-based reset:
-//   If lastRequest < windowStart → reset count to 1 (new window)
-//   Otherwise → increment count
-//   Returns count; if count > RATE_LIMIT_MAX → 429
-//
-// Enforced when: NODE_ENV !== 'test' || BETTER_AUTH_ENFORCE_RATE_LIMIT === '1'
-//
-// KEY TRUST MODEL
-// ---------------
-// The primary rate-limit key is 'invitation:' + req.ip. Express derives req.ip
-// from X-Forwarded-For under app.set('trust proxy', 1) (serverApp.ts). This is
-// safe in the deployed topology because the Passenger/nginx reverse proxy
-// overwrites (not appends) X-Forwarded-For to the real client TCP-socket IP
-// before passing the request to Express. See serverApp.ts for the full
-// DEPLOYMENT CONTRACT comment and the conditions under which this guarantee
-// holds.
-//
-// SECONDARY (TOKEN-SCOPED) KEY — graceful degradation
-// ----------------------------------------------------
-// For the GET /api/auth/invitation/:token lookup endpoint the caller can
-// also pass a secondary key (e.g. the SHA-256 hex of the token). When present,
-// BOTH keys are checked and the stricter limit applies. This means even if
-// req.ip were unreliable (proxy misconfiguration), each unique token still has
-// its own rate-limit bucket and cannot be brute-forced at unbounded rate.
-// ---------------------------------------------------------------------------
-
-const RATE_LIMIT_WINDOW_MS = 60 * 1000; // 60 seconds
-const RATE_LIMIT_MAX = 10;
-
-/**
- * Creates an Express middleware that enforces a per-IP rate limit on invitation
- * routes using the invitationRateLimit table.
- *
- * Skipped when NODE_ENV=test AND BETTER_AUTH_ENFORCE_RATE_LIMIT !== '1',
- * mirroring the existing BETTER_AUTH_ENFORCE_RATE_LIMIT flag (plan.md Pass 12).
- *
- * @param secondaryKeyFn  Optional function to derive a secondary rate-limit key
- *                        from the request (e.g. SHA-256 of the token path param).
- *                        When provided, BOTH the IP-key and the secondary key are
- *                        checked; the first to exceed the limit wins with 429.
- */
-export function invitationRateLimit(pool: Pool, secondaryKeyFn?: (req: Request) => string) {
-  return async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    // Skip in test mode unless enforcement flag is set
-    if (
-      process.env.NODE_ENV === "test" &&
-      process.env.BETTER_AUTH_ENFORCE_RATE_LIMIT !== "1"
-    ) {
-      next();
-      return;
-    }
-
-    const nowMs = Date.now();
-    const windowStart = nowMs - RATE_LIMIT_WINDOW_MS;
-    const ipKey = `invitation:${req.ip ?? "unknown"}`;
-    const secondaryKey = secondaryKeyFn ? secondaryKeyFn(req) : undefined;
-
-    const client = await pool.connect();
-    try {
-      // Prune stale entries from previous windows (same round-trip)
-      await client.query(
-        `DELETE FROM "invitationRateLimit" WHERE "lastRequest" < $1`,
-        [windowStart]
-      );
-
-      // Atomic UPSERT helper: returns the post-increment count for a key
-      const upsertCount = async (k: string): Promise<number> => {
-        const r = await client.query<{ count: number }>(
-          `INSERT INTO "invitationRateLimit" (key, count, "lastRequest")
-           VALUES ($1, 1, $2)
-           ON CONFLICT (key) DO UPDATE SET
-             count = CASE
-               WHEN "invitationRateLimit"."lastRequest" < $3 THEN 1
-               ELSE "invitationRateLimit".count + 1
-             END,
-             "lastRequest" = $2
-           RETURNING count`,
-          [k, nowMs, windowStart]
-        );
-        return r.rows[0]?.count ?? 1;
-      };
-
-      // Check primary (IP-based) key
-      const ipCount = await upsertCount(ipKey);
-      if (ipCount > RATE_LIMIT_MAX) {
-        res.status(429).json({ error: "Too many requests. Please try again later." });
-        return;
-      }
-
-      // Check secondary (token-scoped) key when present — graceful degradation
-      // if req.ip is unreliable due to proxy misconfiguration (see header comment).
-      if (secondaryKey !== undefined) {
-        const secondaryCount = await upsertCount(secondaryKey);
-        if (secondaryCount > RATE_LIMIT_MAX) {
-          res.status(429).json({ error: "Too many requests. Please try again later." });
-          return;
-        }
-      }
-
-      next();
-    } finally {
-      client.release();
-    }
-  };
-}
+import { DEFAULT_BASE_URL } from "../auth/trustedOrigins";
+import { requireSameOrigin } from "../middle/requireSameOrigin";
+import { invitationRateLimit } from "../middle/invitationRateLimit";
 
 // ---------------------------------------------------------------------------
 // Body parser middleware for /api/auth/invitation/accept (plan.md Pass 8)
